+++ Source/SessionDelegate.swift
@@ -225,7 +225,7 @@ extension SessionDelegate: URLSessionDelegate {
             let host = challenge.protectionSpace.host
 
             if
-                let serverTrustEvaluators = session.serverTrustPolicyManager?.serverTrustEvaluators(forHost: host),
+                let serverTrustEvaluators = session.serverTrustManager?.serverTrustEvaluators(forHost: host),
                 let serverTrust = challenge.protectionSpace.serverTrust
             {
                 if serverTrustEvaluators.evaluate(serverTrust, forHost: host) {
